CORE-4720 Updated app_listings/get-app-details format.
Added a groups field that lists all groups the given app is listed under.
Removed the IllegalArgumentException for pipelines.
Refactored component info into a components field that lists info for all components associated with the app or pipeline, and added additional component fields to match the output of metadactyl/list-deployed-components-in-app.